+++ lib/appconfig.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2022
+ * (c) Copyright Ascensio System SIA 2023
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -124,6 +124,13 @@ class AppConfig {
     private $_versionHistory = "versionHistory";
 
     /**
+     * The config key for the protection
+     *
+     * @var string
+     */
+    private $_protection = "protection";
+
+    /**
      * The config key for the chat display setting
      *
      * @var string
@@ -208,6 +215,13 @@ class AppConfig {
     private $_jwtHeader = "jwt_header";
 
     /**
+     * The config key for the allowable leeway in Jwt checks
+     *
+     * @var string
+     */
+    private $_jwtLeeway = "jwt_leeway";
+
+    /**
      * The config key for the settings error
      *
      * @var string
@@ -276,6 +290,13 @@ class AppConfig {
      * @var string
      */
     public $_customization_macros = "customization_macros";
+
+    /**
+     * The config key for the plugins
+     *
+     * @var string
+     */
+    public $_customizationPlugins = "customization_plugins";
 
     /**
      * @param string $AppName - application name
@@ -674,6 +695,30 @@ class AppConfig {
     }
 
     /**
+     * Save protection
+     *
+     * @param bool $value - version history
+     */
+    public function SetProtection($value) {
+        $this->logger->info("Set protection: " . $value, ["app" => $this->appName]);
+
+        $this->config->setAppValue($this->appName, $this->_protection, $value);
+    }
+
+    /**
+     * Get protection
+     *
+     * @return bool
+     */
+    public function GetProtection() {
+        $value = $this->config->getAppValue($this->appName, $this->_protection, "owner");
+        if ($value === "all") {
+            return "all";
+        }
+        return "owner";
+    }
+
+    /**
      * Save chat display setting
      *
      * @param bool $value - display chat
@@ -870,6 +915,26 @@ class AppConfig {
     }
 
     /**
+     * Save plugins setting
+     *
+     * @param bool $value - enable macros
+     */
+    public function SetCustomizationPlugins($value) {
+        $this->logger->info("Set plugins enabled: " . json_encode($value), ["app" => $this->appName]);
+
+        $this->config->setAppValue($this->appName, $this->_customizationPlugins, json_encode($value));
+    }
+
+    /**
+     * Get plugins setting
+     *
+     * @return bool
+     */
+    public function GetCustomizationPlugins() {
+        return $this->config->getAppValue($this->appName, $this->_customizationPlugins, "true") === "true";
+    }
+
+    /**
      * Save the list of groups
      *
      * @param array $groups - the list of groups
@@ -1002,6 +1067,17 @@ class AppConfig {
             $header = "Authorization";
         }
         return $header;
+    }
+
+    /**
+     * Get the Jwt Leeway
+     *
+     * @return int
+     */
+    public function GetJwtLeeway() {
+        $jwtLeeway = (integer)$this->GetSystemValue($this->_jwtLeeway);
+
+        return $jwtLeeway;
     }
 
     /**